// apps/web/lib/mbg/types.ts
export interface ArticleLocation {
  city: string;
  province: string;
}

export interface Article {
  title: string;
  source: string;
  score: number;
  date: string;
  link: string;
  location: ArticleLocation | null;
  dominant_issue: string | null;
}